Convert integration tests into unit tests
The integration tests in the solar-rest project have been converted
to unit tests. Below are the advantages:
 - Unit tests are quicker to run during development.
 - Other developers can clone and build the project as is without the
   need to setup the whole environment. Setting up a MySQL database for
   example could be time consuming.
 - Unit tests are more focused because other layers of the application
   are mocked out. If there is a problem, we know exactly that it is
   caused by the layer of the application we are testing.
Having converted the integration tests of the solar-rest project into
unit tests, there was a need to write unit tests for other layers of
the application. With this approach, each layer can be tested indepen-
dently.
Integration tests will be retroduced later after setting up a
Continuous Integration application such as Jenkins