Add integration test support to the solar-rest project
Integration tests for the solar-rest project will be introduced in a
different project. That project will require classes from the
solar-rest project. Since the solar-rest project is packaged as a war,
we need to deploy it with the classes classifier. The dependent inte-
gration test project will then be able to use the classes in the
solar-rest project by specifying classes specifier in its Maven depen-
dency. Without the classifier, classes defined in a project deployed as
as war file can't be used in another project.