Remove integration testing related code
Our goal in the solar-rest project is to unit test controllers. Each
tests should therefore only test the functionality of the controller
Any integration related code should not be present in unit tests. For
example, while unit testing the PanelController, it is not reasonable
to test the nature of the error sent to the client when a serial number
was not specified during a POST or PUT method. There is one main reason
for this:
 - The controller is not the one in charge of mapping exceptions to
   errors. This is the responsibility of exception handlers. If the ex-
   ception handler didn't do its job correctly, and the unit test
   tested the nature of the JSON error sent back to the client, the
   unit test could fail, indicating that there is a fault in the contro-
   ller, meanwhile the fault is actually in the exception handler. This
   example is therefore more closely related to an integration test
   than a unit test.