Fix the bug in the HourlyElectricityController.delete method
The HourlyElectricityPanelMismatchError error was not caught during
the integration testing of HourlyElectricityController.delete method.
That was due to the fact that the hourly electricity deleted was
fetched from the database and its panel was then set to a new panel
whose Id was that from the request URI. The panel of the existing
hourly electricity was therefore overwrote. The condition to check
whether the panel indicated in the URI and that of the existing hourly
electricity matched, always passed, preventing the HourlyElectricity-
PanelMismatchError error. The guilty here was the HourlyElectricity-
Controller.delete method.